[-] CORE : Code standards [-] CORE : Fixed issue #PSCFI-6314 - "Fatal Error" could appear with Group::getReduction() during the order validation [-] CLASSES : Removed unused getCartCategories() [-] BO : Fixed issue #PSCFI-6310 - CSV Product Import was not working properly if "Product Reference" option was selected [-] FO : Fixed issue #PSCFI-6311 - Number of Product per Page was not working properly for manufacturers and suppliers [-] MO : Fixed several XSS (Kiala, Ogone, Prestassurance, Wexpay) [-] MO : Removed unused and non-existent displaySelectCategoryWidget() function in Prestassurance

+++ classes/PaymentModule.php
@@ -267,7 +267,7 @@ abstract class PaymentModuleCore extends Module
 
 					$group_reduction = (float)GroupReduction::getValueForProduct((int)$product['id_product'], $customer->id_default_group) * 100;
 					if (!$group_reduction)
-						$group_reduction = Group::getReduction((int)$order->id_customer);
+						$group_reduction = (float)Group::getReduction((int)$order->id_customer);
 
 					$quantityDiscount = SpecificPrice::getQuantityDiscount((int)$product['id_product'], Shop::getCurrentShop(), (int)$cart->id_currency, (int)$vat_address->id_country, (int)$customer->id_default_group, (int)$product['cart_quantity']);
 					$unitPrice = Product::getPriceStatic((int)$product['id_product'], true, ($product['id_product_attribute'] ? intval($product['id_product_attribute']) : NULL), 2, NULL, false, true, 1, false, (int)$order->id_customer, NULL, (int)$order->{Configuration::get('PS_TAX_ADDRESS_TYPE')});
